chore: introduce @trezor/dom-utils package
