+++ packages/dom-utils/package.json
@@ -1,0 +1,12 @@
+{
+    "name": "@trezor/dom-utils",
+    "version": "1.0.0",
+    "private": true,
+    "license": "See LICENSE.md in repo root",
+    "sideEffects": false,
+    "main": "src/index",
+    "scripts": {
+        "lint": "eslint '**/*.{ts,tsx,js}'",
+        "type-check": "tsc --build"
+    }
+}
