+++ packages/suite/src/hooks/wallet/useCoinmarketSavingsPaymentInfoCopy.ts
@@ -1,5 +1,5 @@
 import type { SavingsPaymentInfo } from '@suite-services/invityAPI';
-import { copyToClipboard } from '@suite-utils/dom';
+import { copyToClipboard } from '@trezor/dom-utils';
 import { useActions } from '@suite-hooks';
 import * as notificationActions from '@suite-actions/notificationActions';
